Implementation of AWS SSM as secrets manager
